membuat relasi baru, view table, helper format idr, dan memasukkan data

+++ database/migrations/2023_01_22_085057_create_tb_asets_table.php
@@ -25,7 +25,7 @@ class CreateTbAsetsTable extends Migration
             $table->foreignId('umur_ekonomis_id');
             $table->integer('usia_aset');
             $table->foreignId('kondisi_id');
-            $table->decimal('harga');
+            $table->integer('harga');
             $table->timestamps();
         });
     }
